+++ src/pages/admin/components/tableData/TableTopping.js
@@ -1,26 +1,13 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { useMutation, useQuery } from 'react-query';
 import { Link } from 'react-router-dom';
 import convertRupiah from 'rupiah-format';
 import { API, getToppings } from '../../../../config/api';
-import { Modal, Button } from 'react-bootstrap';
+import swal from 'sweetalert';
 
 const TableTopping = () => {
   const { data: toppings, isLoading: loadingTopping, refetch } = useQuery('toppingsCache', getToppings);
-
-  const [idDelete, setIdDelete] = useState(null);
-  const [confirmDelete, setConfirmDelete] = useState(null);
-  const [show, setShow] = useState(false);
-
-  const handleClose = () => setShow(false);
-  const handleShow = () => setShow(true);
-
-  const handlerDelete = (id) => {
-    handleShow();
-    setIdDelete(id);
-  };
-
-  const deleteById = useMutation(async (id) => {
+  const handlerDelete = useMutation(async (id) => {
     try {
       const config = {
         method: 'DELETE',
@@ -28,25 +15,28 @@ const TableTopping = () => {
           Authorization: 'Bearer ' + localStorage.token,
         },
       };
-      const response = await API().delete('/topping/' + id, config);
-      console.log(response);
-      refetch();
+
+      swal({
+        title: 'Are you sure?',
+        text: 'product will be deleted',
+        icon: 'warning',
+        buttons: true,
+        dangerMode: true,
+      }).then(async (willDelete) => {
+        if (willDelete) {
+          const response = await API().delete('/topping/' + id, config);
+          refetch();
+          swal('Product has been deleted!', {
+            icon: 'success',
+          });
+        } else {
+          swal('Product is safe!');
+        }
+      });
     } catch (error) {
       console.log(error);
     }
   });
-
-  const actionDelete = () => {
-    setConfirmDelete(true);
-  };
-
-  useEffect(() => {
-    if (confirmDelete) {
-      handleClose();
-      deleteById.mutate(idDelete);
-      setConfirmDelete(null);
-    }
-  }, [confirmDelete]);
 
   return (
     <>
@@ -86,7 +76,7 @@ const TableTopping = () => {
                       <button
                         className="trash"
                         onClick={() => {
-                          handlerDelete(topping.id);
+                          handlerDelete.mutate(topping.id);
                         }}
                       >
                         <i class="fas fa-trash-alt"></i>
@@ -99,19 +89,6 @@ const TableTopping = () => {
           </table>
         </div>
       </section>
-      <Modal show={show} onHide={handleClose} centered className="text-center modal-delete">
-        <Modal.Body>
-          <h3>Are You Sure Delete?</h3>
-        </Modal.Body>
-        <Modal.Footer className="d-flex justify-content-center">
-          <Button variant="secondary" onClick={handleClose} className="btn-closed">
-            cancel
-          </Button>
-          <Button variant="primary" onClick={actionDelete} className="btn-delete">
-            delete
-          </Button>
-        </Modal.Footer>
-      </Modal>
     </>
   );
 };
